+++ nextcloud_contacts_to_gequdio/__init__.py
@@ -1,0 +1,5 @@
+"""
+Init file for the nextcloud_contacts_to_gequdio package.
+"""
+
+__version__ = "1.1.1"