+++ nextcloud_contacts_to_gequdio/tests/__init__.py
@@ -1,0 +1,3 @@
+"""
+Init file for the tests package of the nextcloud-contacts-to-gequdio project.
+"""